pi/requirements.txt
python-dotenv
wakeonlan